Move ErrorTracker to its own class with a more generic name and add progress tracking
